+++ relive-simulator-core/src/commonMain/kotlin/xyz/qwewqa/relive/simulator/core/stage/strategy/complete/CompleteStrategy.kt
@@ -442,7 +442,7 @@ class CompleteStrategy(val script: CsScriptNode) : Strategy {
         if (cutinUseCounts.getValue(this) > 0) {
           data.cooldown
         } else {
-          (data.startCooldown - actor.cutinInitialCooldownReduction).coerceAtLeast(0)
+          (data.startCooldown - actor.cutinInitialCooldownReduction.toInt()).coerceAtLeast(0)
         }
 
   private fun activateCutin(cutin: BoundCutin) {
